% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amat.R
\name{Amat}
\alias{Amat}
\title{A matrix}
\usage{
Amat(data, chrom = NULL, exclude.marker = NULL)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{chrom}{Only use markers from these chromosomes}

\item{exclude.marker}{Markers to exclude}
}
\value{
A matrix
}
\description{
Calculates the additive (A) relationship matrix from founder genotype probabilities
}
\details{
Additive relationships are calculated from kinship coefficients of order 2 (Gallais 2003). Can be subset to one or more chromosomes (which is useful for the leave-one-chromosome-out kinship method), or specific markers can be excluded after QTL detection.
}
\examples{
\dontrun{
  Amat_example = Amat(data = diallel_example)
  Amat_example = Amat(data = diallel_example, chrom=c(1:11)) #leave chromosome 12 out
  Amat_example = Amat(data = diallel_example, exclude.marker = "solcap_snp_c2_25522")
}

}
\references{
Gallais, A. 2003. Quantitative Genetics and Breeding Methods in  Autopolyploid Plants. Institut National de la Recherche Agronomique, Paris.
}
