% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesCI.R
\name{BayesCI}
\alias{BayesCI}
\title{Bayesian Credible Interval for QTL position}
\usage{
BayesCI(scan1_data, data, chrom, CI.prob = 0.9)
}
\arguments{
\item{scan1_data}{data frame output from scan1}

\item{data}{variable of class \code{\link{diallel_geno_pheno}}}

\item{chrom}{chromosome}

\item{CI.prob}{probability for the credible interval}
}
\value{
first and last marker in the CI
}
\description{
Bayesian Credible Interval for QTL position
}
\details{
Parameter \code{CI.prob} sets the probability for the Bayesian credible interval (e.g., 0.90, 0.95)
using the profile likelihood (posterior mean).
}
\examples{
\dontrun{
  BayesCI(scan1_example, diallel_example, chrom="10", CI.prob=0.9)
  }
}
