% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_onemap.R
\name{convert_onemap}
\alias{convert_onemap}
\title{Generate diaQTL input files from OneMap}
\usage{
convert_onemap(data, digits = 4, conv.bp = 10^6, outstem = "")
}
\arguments{
\item{conv.bp}{value to multiply cM of the map position to have in base pairs}

\item{outstem}{prefix for the pedigree and genotype files for diaQTL}

\item{onemap.object}{onemap_progeny_haplotypes object class}
}
\value{
NULL

#' @examples
\dontrun{
    map <- list(LG1_final, LG2_final)
    progeny_haplot <- onemap::progeny_haplotypes(map,
                                                 most_likely = FALSE,
                                                 ind = "all")
    convert_onemap(progeny_haplot)
}
}
\description{
Generate diaQTL input files from 'onemap_progeny_haplotypes' object class of OneMap R package (version >2.2.0)
}
