% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_polyancestry.R
\name{convert_polyancestry}
\alias{convert_polyancestry}
\title{Create diaQTL input files from polyancestry file}
\usage{
convert_polyancestry(
  filename,
  mapfile = NULL,
  remove.outliers = TRUE,
  outstem = ""
)
}
\arguments{
\item{filename}{Name of polyancestry file}

\item{mapfile}{Optional name of CSV file containing the physical map (marker, chrom, bp)}

\item{remove.outliers}{Should offspring flagged as outliers be removed (default is TRUE)}

\item{outstem}{prefix for output filenames}
}
\description{
Create diaQTL input files from polyancestry file
}
\details{
Creates the pedigree (diaQTL_pedfile.csv) and genotype (diaQTL_genofile.csv) input files needed for \code{\link{read_data}} from the polyancestry output file generated by the PolyOrigin software. PolyOrigin outputs a genetic map in cM. To add a physical map in bp, use the option \code{mapfile}. The input file needed for \code{\link{phased_parents}} (diaQTL_parents.csv) is also created.
}
