% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rabbit.R
\name{convert_rabbit}
\alias{convert_rabbit}
\title{Generate diaQTL input files from RABBIT MagicReconstruct}
\usage{
convert_rabbit(rabbit.outfile, ped.file, outstem)
}
\arguments{
\item{rabbit.outfile}{name of RABBIT output file}

\item{ped.file}{name of RABBIT pedigree file}

\item{outstem}{prefix for the pedigree and genotype files for diaQTL}
}
\description{
Generate diaQTL input files from RABBIT MagicReconstruct
}
