% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diallel_geno_pheno.R
\docType{class}
\name{diallel_geno_pheno-class}
\alias{diallel_geno_pheno-class}
\alias{diallel_geno_pheno}
\title{S4 class with genotype and phenotype data}
\description{
S4 class with genotype and phenotype data
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{Either 2 or 4}

\item{\code{ped}}{data frame with pedigree information. Variables are id,population,mother,father}

\item{\code{map}}{data frame with marker,chrom,and position (either bp or cM)}

\item{\code{geno}}{list of length 2. The first element (named "A") is a list of sparse matrices, one for each marker, with dimensions (indiv x alleles), containing the allele dosages, which are the regression variables for additive effects. The second (optional) element (named "D") has the same structure (list of sparse matrices), but each matrix contains the dosage of allele-pairs, which are the regression variables for digenic dominance effects.}

\item{\code{pheno}}{data frame of phenotypes}

\item{\code{X}}{incidence matrix for fixed effects}

\item{\code{Z}}{incidence matrix for individuals}
}}

