% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diallel_geno_pheno_G.R
\docType{class}
\name{diallel_geno_pheno_G-class}
\alias{diallel_geno_pheno_G-class}
\alias{diallel_geno_pheno_G}
\title{S4 class with genotype and phenotype data and the additive relationship matrix}
\description{
S4 class with genotype and phenotype data and the additive relationship matrix
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{Either 2 or 4}

\item{\code{polyorigin}}{matrix of character strings from the genotype input file}

\item{\code{Xa}}{list of matrices with the expected haplotype dosage (rows) for each parental origin genotype (columns)}

\item{\code{dominance}}{Maximum dosage stored in slot \code{geno}. Integer 1-4 indicating 1 = additive, 2 = digenic dominance, 3 = trigenic dominance, 4 = quadrigenic dominance.}

\item{\code{X.GCA}}{Incidence matrix for GCA effects}

\item{\code{map}}{data frame with marker,chrom, position (cM and/or bp) and bin}

\item{\code{geno}}{list of length equal to the number of marker bins. Each element is a list of length ploidy corresponding to additive, digenic, trigenic, and quadrigenic effects. The elements in the nested list are sparse matrices with dimensions (id x effects).}

\item{\code{pheno}}{data frame of phenotypes}

\item{\code{X}}{incidence matrix for fixed effects}

\item{\code{Z}}{incidence matrix for individuals}

\item{\code{G1}}{additive relationship matrix computed by \code{\link{IBDmat}}}
}}

