% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC_thresh.R
\name{DIC_thresh}
\alias{DIC_thresh}
\title{delta DIC thresholds for scan1}
\usage{
DIC_thresh(genome.size, num.parents, ploidy, alpha = 0.05, dominance = 1)
}
\arguments{
\item{genome.size}{Genome size in Morgans (not centiMorgans)}

\item{num.parents}{Number of parents (2 to 10)}

\item{ploidy}{2 or 4}

\item{alpha}{significance level (0.01, 0.05, 0.10, or 0.20)}

\item{dominance}{1 (additive) or 2 (digenic dominance)}
}
\value{
-deltaDIC threshold
}
\description{
delta DIC thresholds for scan1
}
\details{
Thresholds to control the genome-wide false positive rate at \code{alpha} were determined for half-diallel mating designs with up to 10 parents.
}
\examples{
\dontrun{
  DIC_thresh(genome.size=10, 
             num.parents=4,
             ploidy=4,
             dominance=1,
             alpha=0.05)
  } 
  
}
