% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diplo_freq.R
\name{diplo_freq}
\alias{diplo_freq}
\title{Diplotype frequencies}
\usage{
diplo_freq(data, diplotypes, dosage, position, chrom = NULL)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{diplotypes}{Names of diplotypes}

\item{dosage}{Dosage threshold}

\item{position}{Either "cM" or "bp" for plotting}

\item{chrom}{Names of chromosomes (default is all)}
}
\value{
List containing
\describe{
\item{result}{Data frame with the map and frequency}
\item{plot}{ggplot object}
}
}
\description{
Plot the frequency of individuals with diplotype dosage above a threshold
}
\details{
Useful for visualizing selection in selfed populations.
}
