% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diplo_get.R
\name{diplo_get}
\alias{diplo_get}
\title{Dosage of parental diplotypes}
\usage{
diplo_get(data, marker = NULL, id = NULL)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{marker}{Name of marker}

\item{id}{Name of individual}
}
\value{
Matrix of (id or markers) x parental diplotypes
}
\description{
Dosage of parental diplotypes
}
\details{
Function can be used to get parental diplotype dosage estimates at a single marker for all individuals (in which case \code{id} should be \code{NULL}) or for a single individual for all markers (in which case \code{marker} should be \code{NULL})
}
\examples{
\dontrun{
  diplo_example = diplo_get(data = diallel_example, 
                          marker = "solcap_snp_c2_25522")
  diplo_example = diplo_get(data = diallel_example, 
                          id = "W15263-8R")
}

}
