% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1codes.R
\docType{data}
\name{F1codes}
\alias{F1codes}
\title{Genotype codes for F1 populations}
\format{
character vector
}
\usage{
data(F1codes)
}
\description{
Character vector with the 100 possible tetraploid genotypes for a F1 population. Maternal haplotypes are denoted 1,2,3,4 and paternal haplotypes 5,6,7,8.
}
\keyword{datasets}
