% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faa.R
\name{faa}
\alias{faa}
\title{Construct incidence matrix for additive x additive epistasis}
\usage{
faa(j, k, data)
}
\arguments{
\item{j}{first marker}

\item{k}{second marker}

\item{data}{variable inheriting from class \code{\link{diallel_geno}}}
}
\value{
Matrix with dimensions nind x (n.hap)^2
}
\description{
Construct incidence matrix for additive x additive epistasis
}
\keyword{internal}
