% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_map.R
\name{fine_map}
\alias{fine_map}
\title{Visualize haplotype switches for fine mapping}
\usage{
fine_map(data, haplotype, interval, trait = NULL)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{haplotype}{Name of parental haplotype}

\item{interval}{2-vector with marker names}

\item{trait}{Name of trait to plot (optional)}
}
\value{
ggplot2 variable
}
\description{
Visualize haplotype switches for fine mapping
}
\details{
Function returns graphic for all individuals with a haplotype switch (defined as change in dosage from 0 to \eqn{\geq} 1 or vice versa) for \code{haplotype} within \code{interval}. If \code{trait} is included, the trait values for each individual are displayed on the right side. The function requires map positions in bp to be included in \code{data}.
}
