% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Get map summary from diallel_geno object}
\usage{
get_map(data, summary = TRUE)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{summary}{logical, if TRUE (default) returns total sizes per chromosome, if FALSE returns the map}
}
\value{
data frame with map summary or the map
}
\description{
Get map summary from diallel_geno object
}
\examples{
\dontrun{
  get_map(diallel_example)
}

}
