% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gprep.R
\name{Gprep}
\alias{Gprep}
\title{Compute realized relationship matrix for fitQTL}
\usage{
Gprep(data, marker)
}
\arguments{
\item{data}{variable inheriting from class \code{\link{diallel_geno_pheno}}}

\item{marker}{markers that will be used for \code{\link{fitQTL}}}
}
\value{
variable inheriting from class \code{\link{diallel_geno_pheno_G}}
}
\description{
Compute realized relationship matrix for fitQTL
}
\details{
Computes the additive realized relationship matrix with \code{\link{IBDmat}} so that \code{\link{fitQTL}} does not need to and can run more quickly. The chromosome(s) containing \code{marker} is (are) excluded from the relationship calculation.
}
\examples{
\dontrun{
  diallel_example <- Gprep(data = diallel_example, marker = "solcap_snp_c2_25522")
}

}
