% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_freq.R
\name{haplo_freq}
\alias{haplo_freq}
\title{Haplotype frequencies}
\usage{
haplo_freq(
  data,
  haplotypes,
  dosage,
  id = NULL,
  position = "cM",
  chrom = NULL,
  markers = NULL
)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{haplotypes}{Names of haplotypes}

\item{dosage}{Dosage threshold}

\item{id}{Vector of id names (default is entire population)}

\item{position}{Either "cM" (default) or "bp" for plotting}

\item{chrom}{Names of chromosomes (default is all)}

\item{markers}{Optional, markers to indicate with dashed line. Only available when plotting a single chromosome.}
}
\value{
List containing
\describe{
\item{result}{Data frame with the map and frequency}
\item{plot}{ggplot object}
}
}
\description{
Plots the frequency of individuals with haplotype dosage above a threshold
}
\details{
Useful for visualizing selection in selfed populations. For multiple chromosomes, each haplotype is shown in its own panel using facet_wrap. For one chromosome, the haplotypes are shown on the same set of axes.
}
