% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_match.R
\name{haplo_match}
\alias{haplo_match}
\title{Match up parental haplotypes}
\usage{
haplo_match(file1, file2, chrom)
}
\arguments{
\item{file1}{name of CSV phased parent file}

\item{file2}{name of CSV phased parent file}

\item{chrom}{chromosome}
}
\value{
Data frame with results
}
\description{
Match up parental haplotypes
}
\details{
Designed to match up parental haplotypes between two phased parent files based on genetic distance. In the plots, the haplotypes in file1 are numbered 1-4, and those in file2 are numbered 5-8.
}
