% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_plot.R
\name{haplo_plot}
\alias{haplo_plot}
\title{Plot parental haplotype dosage}
\usage{
haplo_plot(data, id, chrom, position = "cM", markers = NULL)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{id}{Name of individual(s)}

\item{chrom}{Name of chromosome}

\item{position}{Either "cM" (default) or "bp"}

\item{markers}{Optional, markers to indicate with dashed line}
}
\value{
ggplot object
}
\description{
Plot parental haplotype dosages across the chromosome for one individual
}
\details{
For "cM" plotting, only one marker per bin is displayed. For "bp" plotting, all markers are included. If multiple individuals are included in \code{id}, then the plot shows the probability that a haplotype is present in all individuals.
}
\examples{
\dontrun{
haplo_plot(data = diallel_example, 
            id = "W15263-8R", 
            chrom = 10)
            
haplo_plot(data = diallel_example, 
            id = "W15263-8R", 
            chrom = 10,
            marker = "solcap_snp_c2_25522")
}
}
