% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_switch.R
\name{haplo_switch}
\alias{haplo_switch}
\title{Find haplotype switches}
\usage{
haplo_switch(data, marker, haplotype, position, jump = 0.8)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{marker}{Name of focal marker}

\item{haplotype}{Name of parental haplotype}

\item{position}{Either "cM" or "bp"}

\item{jump}{Change in dosage to designate haplotype switch}
}
\value{
Data frame with the locations of the nearest haplotype switch on the left and right side of the focal marker for all individuals with a haplotype switch
}
\description{
Find haplotype switches
}
\details{
Designed to help with fine mapping of QTL. Function returns the location of the nearest haplotype switch on both sides (left, right) of a marker and haplotype of interest that represent the presumed location of a QTL allele.
}
