% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_unique.R
\name{haplo_unique}
\alias{haplo_unique}
\title{Find interval to differentiate haplotypes}
\usage{
haplo_unique(filename, marker, haplotypes)
}
\arguments{
\item{filename}{Name of CSV input file}

\item{marker}{Name of target marker}

\item{haplotypes}{Vector of haplotype names to differentiate}
}
\value{
List containing
\describe{
\item{haplo}{Data frame of haplotypes}
\item{dendro}{Dendrogram from hclust(method="average")}
}
}
\description{
Finds the smallest interval such that the phased SNP haplotypes are all unique
}
\details{
The input file (diaQTL_parents.csv) should be generated by \code{\link{read_polyancestry}}. Results can be visualized using \code{\link{phased_parents}}.
}
