% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{haplotypes}
\alias{haplotypes}
\title{Dosage of parental haplotypes}
\usage{
haplotypes(data, marker = NULL, id = NULL)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{marker}{Name of marker}

\item{id}{Name of individual}
}
\value{
Matrix of (id or markers) x parental haplotypes
}
\description{
Dosage of parental haplotypes
}
\details{
Function can be used to get parental haplotype dosage estimates at a single marker for all individuals (in which case \code{id} should be \code{NULL}) or for a single individual for all markers (in which case \code{marker} should be \code{NULL})
}
\examples{
\dontrun{
  haplo_example = haplotypes(data = diallel_example, 
                          marker = "solcap_snp_c2_25522")
  haplo_example = haplotypes(data = diallel_example, 
                          id = "W15263-8R")
}

}
