% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOD_thresh.R
\name{LOD_thresh}
\alias{LOD_thresh}
\title{LOD thresholds for scan1}
\usage{
LOD_thresh(genome.size, num.parents, ploidy, alpha = 0.05, dominance = 1)
}
\arguments{
\item{genome.size}{Genome size in Morgans (not centiMorgans)}

\item{num.parents}{Number of parents}

\item{ploidy}{2 or 4}

\item{alpha}{false positive rate: 0.01, 0.05, 0.10, or 0.20}

\item{dominance}{1 (additive) or 2 (digenic dominance)}
}
\value{
LOD threshold
}
\description{
LOD thresholds for scan1
}
\details{
LOD thresholds to control the genome-wide false positive rate at 0.05 were determined via simulation for up to 20 parents and genome sizes up to 12 Morgans. A monotone increasing concave curve was fit to these results using R package \code{scam} and is used for prediction. (The LOD threshold does not depend on population size.)
}
\examples{
\dontrun{
  LOD_thresh(genome.size=10, 
            num.parents=4,
            ploidy=4,
            dominance=1,
            alpha=0.05)
  } 
  
}
