% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LODthresh.R
\name{LODthresh}
\alias{LODthresh}
\title{LOD thresholds for scan1}
\usage{
LODthresh(genome.size, num.parents, ploidy)
}
\arguments{
\item{genome.size}{Genome size in Morgans (not centiMorgans)}

\item{num.parents}{Number of parents}

\item{ploidy}{2 or 4}
}
\value{
LOD threshold
}
\description{
LOD thresholds for scan1
}
\details{
LOD thresholds to control the genome-wide false positive rate at 0.05 were determined via simulation for up to 20 parents and genome sizes up to 12 Morgans. A monotone increasing concave curve was fit to these results using R package \code{scam} and is used for prediction. (The LOD threshold does not depend on population size.)
}
