% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased_parents.R
\name{phased_parents}
\alias{phased_parents}
\title{Visualize phased SNPs of parents}
\usage{
phased_parents(filename, interval, markers, parents)
}
\arguments{
\item{filename}{Name of CSV input file}

\item{interval}{Vector of length 2 with the first and last marker names}

\item{markers}{Vector of marker names to plot}

\item{parents}{Vector of parent names to plot}
}
\value{
ggplot2 object
}
\description{
Visualize phased SNPs of parents
}
\details{
The solid circles in the figure represent the allele counted by dosage.
}
