% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dosage.R
\name{plot_dosage}
\alias{plot_dosage}
\title{Plot parental allele dosage}
\usage{
plot_dosage(data, id, chrom, distance, marker = NULL)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{id}{Name of individual}

\item{chrom}{Name of chromosome}

\item{distance}{Either "cM" for centiMorgans or "bp" for base pairs (depending on position used in the input map)}

\item{marker}{Optional, position of marker indicated with dashed line}
}
\value{
ggplot object
}
\description{
Plot parental allele dosages across the chromosome for one individual
}
\examples{
\dontrun{
plot_dosage(data = diallel_example, 
            id = "W15263-8R", 
            chrom = 10)
            
plot_dosage(data = diallel_example, 
            id = "W15263-8R", 
            chrom = 10,
            marker = "solcap_snp_c2_25522")
}
}
