% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polyancestry.R
\name{read_polyancestry}
\alias{read_polyancestry}
\title{Create diaQTL input files from polyancestry file}
\usage{
read_polyancestry(filename, remove.outliers = TRUE)
}
\arguments{
\item{filename}{Name of polyancestry file}

\item{remove.outliers}{Should offspring flagged as outliers be removed (default is TRUE)}
}
\description{
Create diaQTL input files from polyancestry file
}
\details{
Creates the pedigree (diaQTL_pedfile.csv) and genotype (diaQTL_genofile.csv) input files needed for \code{\link{read_data}} from the polyancestry output file generated by the PolyOrigin software. The input file (diaQTL_valents.csv) needed for \code{\link{valents_summary}} is also created.
}
