% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBGLR.R
\name{runBGLR}
\alias{runBGLR}
\title{Xqtl is list of length equal to the number of QTL, and each element is a list of maximum length 4, for additive up to quadrigenic dominance. Xaa is list of matrices for epistasis.}
\usage{
runBGLR(
  params,
  y,
  Xfix,
  Xgca = NULL,
  Xqtl = NULL,
  Xaa = NULL,
  polyG = NULL,
  saveEffects
)
}
\description{
Xqtl is list of length equal to the number of QTL, and each element is a list of maximum length 4, for additive up to quadrigenic dominance. Xaa is list of matrices for epistasis.
}
\keyword{internal}
