% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S1codes.R
\docType{data}
\name{S1codes}
\alias{S1codes}
\title{Genotype codes for S1 populations}
\format{
character vector
}
\usage{
data(S1codes)
}
\description{
Character vector with the 35 possible tetraploid genotypes for a S1 population. Haplotypes are denoted 1,2,3,4.
}
\keyword{datasets}
