% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1_summary.R
\name{scan1_summary}
\alias{scan1_summary}
\title{Summary of scan1 result}
\usage{
scan1_summary(
  scan1_data,
  display = T,
  thresh = NULL,
  chromosome = NULL,
  distance = "cM"
)
}
\arguments{
\item{scan1_data}{output from scan1}

\item{display}{Logical variable whether to plot the LOD score}

\item{thresh}{optional, LOD threshold for plotting}

\item{chromosome}{string with chrom name(s) to plot. By default, all chromosomes are plotted}

\item{distance}{string with "cM" for centiMorgans or "bp" for basepairs}
}
\value{
List containing
\describe{
\item{peaks}{Data frame of the markers with the highest LOD score per chromosome}
\item{plot}{ggplot object}
}
}
\description{
Summary of scan1 result
}
\examples{
\dontrun{
  scan1_summary( scan1_example )
  scan1_summary( scan1_example, chromosome = "10" )
  scan1_summary( scan1_example, chromosome = c( "10", "12" ) ) 
  }
}
