% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{set_params}
\alias{set_params}
\title{Determine parameters for scan1}
\usage{
set_params(data, trait, dominance = 1, tol = 0.1, burnIn = 50, nIter = 1000)
}
\arguments{
\item{data}{Variable of class \code{\link{diallel_geno_pheno}}}

\item{trait}{Name of trait}

\item{dominance}{Complexity of the genetic model (1-4). See Details.}

\item{tol}{tolerance for estimating the median}

\item{burnIn}{initial value for burnIn parameter}

\item{nIter}{initial value for nIter parameter}
}
\value{
List containing
\describe{
\item{burnIn}{Number of burn-in iterations}
\item{nIter}{Total number of iterations}
}
}
\description{
Determine parameters for scan1
}
\details{
The burn-in and total number of iterations are determined using the Raftery and Lewis diagnostic from R package \code{coda}, based on a 95\% probability that the estimated median of the additive effects is between the quantiles (0.5-tol) to (0.5+tol). For greater precision, decrease the \code{tol} parameter. Parameter \code{dominance} controls the complexity of the genetic model: 1 = additive, 2 = digenic dominance, 3 = trigenic dominance, 4 = quadrigenic dominance.
}
\examples{
\dontrun{
  par1 <- set_params(data = diallel_example,
                     trait = "tuber_shape")
}
  
}
