% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valents_summary.R
\name{valents_summary}
\alias{valents_summary}
\title{Summarize valents results}
\usage{
valents_summary(filename)
}
\arguments{
\item{filename}{Name of valents input file}
}
\value{
List containing
\describe{
\item{result}{data frame with results}
\item{plot}{stacked bar chart from ggplot2}
}
}
\description{
Tabulate the proportion of bivalents and quadrivalents from PolyOrigin
}
\details{
The valents input file should be generated with \code{\link{read_polyancestry}}.
}
