% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amat.R
\name{Amat}
\alias{Amat}
\title{A matrix}
\usage{
Amat(data, chrom = NULL)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{chrom}{Optional, vector of chromosome names to include}
}
\value{
A matrix
}
\description{
Calculates the additive (A) relationship matrix from founder genotype probabilities
}
\details{
Additive relationships are calculated from kinship coefficients of order 2 (Gallais 2003). Can specify to use only a subset of the chromosomes (by default, all chromosomes are used). Calculated based on the marker bins.
}
\examples{
\dontrun{
  Amat_example = Amat(data = diallel_example)
  Amat_example = Amat(data = diallel_example, chrom=c(1:11)) #leave chromosome 12 out
}

}
\references{
Gallais, A. 2003. Quantitative Genetics and Breeding Methods in  Autopolyploid Plants. Institut National de la Recherche Agronomique, Paris.
}
