% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diallel_geno.R
\docType{class}
\name{diallel_geno-class}
\alias{diallel_geno-class}
\alias{diallel_geno}
\title{S4 class with genotype data}
\description{
S4 class with genotype data
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{Either 2 or 4}

\item{\code{dominance}}{Integer 1-4 indicating 1 = additive, 2 = digenic dominance, 3 = trigenic dominance, 4 = quadrigenic dominance.}

\item{\code{X.GCA}}{Incidence matrix for GCA effects}

\item{\code{map}}{data frame with marker,chrom, position (cM and/or bp) and bin}

\item{\code{geno}}{list of length equal to the number of marker bins. Each element is a list of length \code{dominance}. The elements in the nested list are sparse matrices with dimensions (id x effects).}
}}

