% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIC_thresh.R
\name{DIC_thresh}
\alias{DIC_thresh}
\title{delta DIC thresholds for scan1}
\usage{
DIC_thresh(genome.size, num.parents, ploidy, alpha = 0.05, dominance = 1)
}
\arguments{
\item{genome.size}{Genome size in Morgans (not centiMorgans)}

\item{num.parents}{Number of parents}

\item{ploidy}{2 or 4}

\item{alpha}{false positive rate: 0.01, 0.05, 0.10, or 0.20}

\item{dominance}{1 (additive) or 2 (digenic dominance)}
}
\value{
deltaDIC threshold
}
\description{
delta DIC thresholds for scan1
}
\details{
delta DIC thresholds to control the genome-wide false positive rate at alpha-level were determined via simulation for up to 20 parents and genome sizes up to 12 Morgans. A monotone increasing concave curve was fit to these results using R package \code{scam} and is used for prediction. (The LOD threshold does not depend on population size.)
}
\examples{
\dontrun{
  DIC_thresh(genome.size=10, 
             num.parents=4,
             ploidy=4,
             dominance=1,
             alpha=0.05)
  } 
  
}
