% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dmat.R
\name{Dmat}
\alias{Dmat}
\title{D matrix}
\usage{
Dmat(data, chrom = NULL, exclude.marker = NULL)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{chrom}{Only use markers from these chromosomes}

\item{exclude.marker}{Markers to exclude}
}
\value{
Digenic dominance relationship matrix
}
\description{
Calculates the dominance (D) relationship matrix from founder genotype probabilities
}
\details{
Dominance relationships are calculated from kinship coefficients of order 4 (Gallais 2003). An entire chromosome or specific markers can be excluded.
}
\examples{
\dontrun{
  Dmat_example = Dmat(data = diallel_example)
  Dmat_example = Dmat(data = diallel_example, chrom=c(10,11)) #leave chromosome 12 out
  Dmat_example = Dmat(data = diallel_example, exclude.marker = "solcap_snp_c2_25522")
}


}
\references{
Gallais, A. 2003. Quantitative Genetics and Breeding Methods in  Autopolyploid Plants. Institut National de la Recherche Agronomique, Paris.
}
