% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQTL.R
\name{fitQTL}
\alias{fitQTL}
\title{Fit a single QTL model}
\usage{
fitQTL(
  data,
  trait,
  marker,
  params,
  dominance = 1,
  cofactor = NULL,
  CI.prob = 0.9,
  polygenic = TRUE
)
}
\arguments{
\item{data}{variable of class \code{\link{diallel_geno_pheno}}}

\item{trait}{name of trait}

\item{marker}{name of marker to fit as QTL}

\item{params}{list containing the number of burn-in (burnIn) and total iterations (nIter)}

\item{dominance}{dominance degree}

\item{cofactor}{optional, see Details for format.}

\item{CI.prob}{probability for Bayesian credible interval}

\item{polygenic}{TRUE/FALSE whether to include a polygenic effect}
}
\value{
List containing
\describe{
\item{r2}{sauared correlation betwen fitted and observed values}
\item{deltaDIC}{Deviance Information Criterion relative to null model}
\item{resid}{Residuals}
\item{var}{Matrix with proportion of variance for the effects}
\item{effects}{List of matrices containing the additive and higher order effects}
\item{plots}{List of ggplot objects for the effects}
}
}
\description{
Fit a single QTL model
}
\details{
The number of burn-in and total iterations in \code{params} can be estimated using \code{\link{set_params}}. Parameter \code{dominance} controls the genetic model for the QTL: 1 = additive, 2 = digenic dominance, 3 = trigenic dominance, 4 = quadrigenic dominance. The optional argument \code{cofactor} should be a list with three components: marker = name of the marker; dominance = 1, 2, 3, or 4; epistasis = TRUE/FALSE. When \code{polygenic = TRUE}, the model includes a random effect with covariance equal to the additive relationship computed by \code{\link{IBDmat}}, leaving out chromosome(s) with the QTL and cofactor (if present). Parameter \code{CI.prob} sets the probability (e.g., 0.90, 0.95) for the Bayesian credible interval for the estimated effects (to disable plotting of the CI, use \code{CI.prob=NULL}).

The LOD and deltaDIC values returned by the function are relative to a model without \code{marker} but including the cofactor and polygenic effect when present. If \code{polygenic = FALSE}, the null model includes a GCA effect. r2 is the squared correlation between the fitted and observed values. The returned list \code{effects} contains the additive (and when included) digenic dominance effects. The proportion of variance for each effect is returned in \code{var}. The returned object \code{plots$dom} shows the digenic dominance effects above the diagonal, and below the diagonal is the sum of the additive and digenic dominance effects.
}
\examples{
\dontrun{
## additive effects
params1 <- set_params( diallel_example, trait = "tuber_shape" ,q=0.05,r=0.05)

fit1 <- fitQTL( data = diallel_example, 
                 trait = "tuber_shape", 
                 params = params1, 
                 marker = "solcap_snp_c2_25522",
                 CI.prob = 0.9)
                 
## additive + dominance effects
params2 <- set_params( diallel_example, trait = "tuber_shape", dominance=2,q=0.05,r=0.05)

fit2 <- fitQTL( data = diallel_example, 
                 trait = "tuber_shape", 
                 params = params2, 
                 marker = "solcap_snp_c2_25522",
                 dominance = 2,
                 CI.prob=0.9)
                 
}
                 
}
