% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQTL.R
\name{fitQTL}
\alias{fitQTL}
\title{Fit a single QTL model}
\usage{
fitQTL(
  data,
  trait,
  marker,
  params,
  dominance = 1,
  CI.prob = 0.9,
  polygenic = TRUE,
  cofactor = NULL
)
}
\arguments{
\item{data}{Variable of class \code{\link{diallel_geno_pheno}}}

\item{trait}{Name of trait}

\item{marker}{Name of marker to fit as QTL}

\item{params}{List containing the number of burn-in (burnIn) and total iterations (nIter)}

\item{dominance}{Dominance degree}

\item{CI.prob}{Probability for Bayesian credible interval}

\item{polygenic}{TRUE/FALSE whether to include polygenic background effect}

\item{cofactor}{Name of marker to fit as cofactor (optional)}
}
\value{
List containing
\describe{
\item{r2}{sauared correlation betwen fitted and observed values}
\item{deltaDIC}{Deviance Information Criterion relative to null model}
\item{resid}{Residuals}
\item{var}{Matrix with proportion of variance for the effects}
\item{effects}{List of matrices containing the additive and higher order effects}
\item{plots}{List of ggplot objects for the effects}
}
}
\description{
Fit a single QTL model
}
\details{
LOD score is the difference between the log10-likelihood for the QTL model vs. no QTL model (higher is better). deltaDIC is the difference between the Deviance Information Criterion for the QTL model vs. no QTL model (lower values is better). r2 is the squared correlation between the fitted and observed values. Parameter \code{dominance} controls the genetic model: 1 = additive, 2 = digenic dominance, 3 = trigenic dominance, 4 = quadrigenic dominance. MCMC \code{params} can be estimated using \code{\link{set_params}}. Parameter \code{CI.prob} sets the probability (e.g., 0.90, 0.95) for the Bayesian credible interval for the estimated effects (to disable plotting of the CI, use \code{CI.prob=NULL}). The returned list \code{effects} contains the additive (and when included) digenic dominance effects. The proportion of variance for each effect is returned in \code{var}. The returned object \code{plots$dom} shows the digenic dominance effects above the diagonal, and below the diagonal is the sum of the additive and digenic dominance effects. The polygenic background effect has covariance equal to the additive relationship computed by \code{\link{IBDmat}}, leaving out the chromosome with the QTL. For faster execution with the polygenic model, use \code{\link{Gprep}} first.
}
\examples{
\dontrun{
## additive effects
params1 <- set_params( diallel_example, trait = "tuber_shape" ,q=0.05,r=0.05)

fit1 <- fitQTL( data = diallel_example, 
                 trait = "tuber_shape", 
                 params = params1, 
                 marker = "solcap_snp_c2_25522",
                 CI.prob = 0.9)
                 
## additive + dominance effects
params2 <- set_params( diallel_example, trait = "tuber_shape", dominance=2,q=0.05,r=0.05)

fit2 <- fitQTL( data = diallel_example, 
                 trait = "tuber_shape", 
                 params = params2, 
                 marker = "solcap_snp_c2_25522",
                 dominance = 2,
                 CI.prob=0.9)
                 
}
                 
}
