% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitQTL.R
\name{fitQTL}
\alias{fitQTL}
\title{Fit multiple QTL model}
\usage{
fitQTL(
  data,
  trait,
  qtl,
  epistasis = NULL,
  polygenic = FALSE,
  params = list(burnIn = 100, nIter = 5000),
  CI.prob = 0.9
)
}
\arguments{
\item{data}{variable of class \code{\link{diallel_geno_pheno}}}

\item{trait}{name of trait}

\item{qtl}{data frame, see Details}

\item{epistasis}{optional data frame, see Details}

\item{polygenic}{TRUE/FALSE whether to include additive polygenic effect}

\item{params}{list containing the number of burn-in (burnIn) and total iterations (nIter)}

\item{CI.prob}{probability for Bayesian credible interval}
}
\value{
List containing
\describe{
\item{deltaDIC}{DIC relative to model with GCA but no QTL effects}
\item{resid}{residuals}
\item{var}{matrix with proportion of variance for the effects}
\item{effects}{list with two matrices, \code{additive} and \code{digenic}, with markers on the rows and effects on the columns}
\item{plots}{list of ggplot objects, one for each marker, containing elements \code{additive} and \code{digenic}. The digenic plot has digenic effects above the diagonal and the sum of additive and digenic effects below the diagonal.}
}
}
\description{
Fit multiple QTL model
}
\details{
Argument \code{qtl} is a data frame with columns \code{marker} and \code{dominance}
to specify the marker name and highest order effect (1 = additive, 2 = digenic dominance,
3 = trigenic dominance, 4 = quadrigenic dominance). All effects up to the value in \code{dominance}
are included.  Optional argument \code{epistasis} is a data frame with columns \code{marker1} and
\code{marker2}, where each row specifies an additive x additive epistatic interaction.
The number of burn-in and total iterations in \code{params} can be estimated using
\code{\link{set_params}}. Parameter \code{CI.prob} sets the probability (e.g., 0.90, 0.95)
for the Bayesian credible interval for the estimated effects (to disable plotting of the CI,
use \code{CI.prob=NULL}).
}
\examples{
\dontrun{
## getting minimum burnIn and nIter for one qtl
set_params(data = diallel_example, 
           trait = "tuber_shape", 
           q = 0.05, 
           r = 0.025, 
           qtl = data.frame(marker="solcap_snp_c2_25522",dominance=2),
           polygenic = TRUE)
           
## additive effects
fit1 <- fitQTL(data = diallel_example, 
               trait = "tuber_shape", 
               params = list(burnIn=100,nIter=5000), 
               qtl = data.frame(marker="solcap_snp_c2_25522",dominance=1),
               CI.prob = 0.9)

## additive + digenic dominance effects                            
fit2 <- fitQTL(data = diallel_example, 
               trait = "tuber_shape", 
               params = list(burnIn=100,nIter=5000), 
               qtl = data.frame(marker="solcap_snp_c2_25522",dominance=2),
               CI.prob=0.9)
               
## getting minimum burnIn and nIter for two qtl with epistasis
set_params(data = diallel_example, 
           trait = "tuber_shape", 
           q = 0.05, 
           r = 0.025, 
           qtl = data.frame(marker=c("PotVar0099535","solcap_snp_c2_25522"),
                            dominance=c(2,1)),
           epistasis = data.frame(marker1="solcap_snp_c2_25522",marker2="PotVar0099535"),
           polygenic = TRUE)
           
## additive + digenic dominance effects for both QTL
fit3 <- fitQTL(data = diallel_example, trait = "tuber_shape", 
               params = list(burnIn=100,nIter=5000),
               qtl = data.frame(marker=c("PotVar0099535","solcap_snp_c2_25522"),
                                dominance=c(2,2)), 
               polygenic = TRUE, CI.prob = 0.9)
               
## additive + digenic dominance effects for both QTL + their epistatic effects
fit4 <- fitQTL(data = diallel_example, trait = "tuber_shape", 
               params = list(burnIn=100,nIter=5000),
               qtl = data.frame(marker=c("PotVar0099535","solcap_snp_c2_25522"),
                                dominance=c(2,2)), 
               epistasis = data.frame(marker1="solcap_snp_c2_25522",marker2="PotVar0099535"),
               polygenic = TRUE, CI.prob = 0.9)
               
## additive + digenic dominance effects for three QTL + all their epistatic effects
fit5 <- fitQTL(data = diallel_example, trait = "tuber_shape", 
               params = list(burnIn=100,nIter=5000),
               qtl = data.frame(marker=c("PotVar0099535",
                                         "solcap_snp_c1_6427",
                                         "solcap_snp_c2_25522"),
                                dominance=c(2,2,2)), 
               epistasis = data.frame(marker1=c("solcap_snp_c2_25522",
                                                "solcap_snp_c2_25522",
                                                "PotVar0099535"),
                                      marker2=c("PotVar0099535",
                                                "solcap_snp_c1_6427",
                                                "solcap_snp_c1_6427")),
               polygenic = TRUE, CI.prob = 0.9)

}
                 
}
