% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_cluster.R
\name{haplo_cluster}
\alias{haplo_cluster}
\title{Cluster parental haplotypes}
\usage{
haplo_cluster(filename, marker, haplotypes = NULL)
}
\arguments{
\item{filename}{Name of CSV input file}

\item{marker}{Either target marker or marker interval (see Details).}

\item{haplotypes}{Vector of haplotype names (default is all)}
}
\value{
List containing
\describe{
\item{haplo}{Data frame of haplotypes}
\item{dendro}{Dendrogram}
}
}
\description{
Cluster parental haplotypes
}
\details{
The input file (diaQTL_parents.csv) should be generated by \code{\link{read_polyancestry}}. The argument \code{marker} can be either a single marker or vector of two markers. If a single marker, the function finds the smallest interval containing that marker such that the phased SNP haplotypes are all unique. If two markers are provided, that interval is used. Clustering utilizes hclust(method="average"). See also \code{\link{phased_parents}} for an additional visualization tool.
}
