% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_freq.R
\name{haplo_freq}
\alias{haplo_freq}
\title{Frequency of haplotype dosages}
\usage{
haplo_freq(data, haplotypes, dosage, position, chrom = NULL)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{haplotypes}{Names of haplotypes}

\item{dosage}{Dosage threshold}

\item{position}{Either "cM" or "bp" for plotting}

\item{chrom}{Names of chromosomes (default is all)}
}
\value{
List containing
\describe{
\item{result}{Data frame with the map and frequency}
\item{plot}{ggplot object}
}
}
\description{
Plot the frequency of individuals with haplotype dosage above a threshold
}
\details{
Useful for visualizing selection in selfed populations.
}
