% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDmat.R
\name{IBDmat}
\alias{IBDmat}
\title{Realized IBD relationship}
\usage{
IBDmat(data, dominance = 1, chrom = NULL, n.core = 1)
}
\arguments{
\item{data}{Variable inheriting from class \code{\link{diallel_geno}}}

\item{dominance}{One of 1,2,3,4}

\item{chrom}{Optional, vector of chromosome names to include}

\item{n.core}{number of cores for parallel execution}
}
\value{
Relationship matrix
}
\description{
Calculates realized relationship matrices (additive and dominance) from founder genotype probabilities
}
\details{
Parameter \code{dominance} refers to 1 = additive, 2 = digenic, 3 = trigenic, 4 = quadrigenic (Gallais 2003).  Can specify to use only a subset of the chromosomes (by default, all chromosomes are used). Calculated based on the marker bins.
}
\examples{
\dontrun{
  IBD_example = IBDmat(data = diallel_example, dominance=1) #additive
  IBD_example = IBDmat(data = diallel_example, dominance=2) #digenic dominance
}


}
\references{
Gallais, A. 2003. Quantitative Genetics and Breeding Methods in  Autopolyploid Plants. Institut National de la Recherche Agronomique, Paris.
}
