% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1.R
\name{scan1}
\alias{scan1}
\title{Single QTL scan}
\usage{
scan1(
  data,
  trait,
  params,
  dominance = 1,
  cofactor = NULL,
  chrom = NULL,
  n.core = 1
)
}
\arguments{
\item{data}{variable of class \code{\link{diallel_geno_pheno}}}

\item{trait}{name of trait}

\item{params}{list containing burnIn and nIter}

\item{dominance}{maximum dominance for the scan, see Details}

\item{cofactor}{optional data frame, see Details}

\item{chrom}{names of chromosomes to scan (default is all)}

\item{n.core}{number of cores for parallel execution}
}
\value{
Data frame containing the map, LL, and deltaDIC.
}
\description{
Performs a linear regression for each position in the map.
}
\details{
Parameter \code{dominance} has possible values of 1 = additive, 2 = digenic dominance, 3 = trigenic dominance, 4 = quadrigenic dominance. MCMC \code{params} can be estimated using \code{\link{set_params}}. Optional argument \code{cofactor} is used to include other markers in the model during the scan, which can improve statistical power with multiple QTL. It is a data frame with three columns: marker = name of the marker, dominance = 1 to 4, and epistasis = TRUE/FALSE. Function returns deltaDIC = DIC for the QTL model relative to null model with only GCA effects for the parents, as well as LL = posterior mean of the log-likelihood, which is used by \code{\link{BayesCI}}.
}
\examples{
\dontrun{
  par1 <- set_params(data = diallel_example,
                     trait = "tuber_shape")
                      
  scan1_example <- scan1(data = diallel_example,
                chrom = 10,
                trait = "tuber_shape",
                params = par1)
}

}
