% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1.R
\name{scan1}
\alias{scan1}
\title{Single QTL scan}
\usage{
scan1(
  data,
  trait,
  params,
  dominance = 1,
  chrom = NULL,
  cofactor = NULL,
  n.core = 1
)
}
\arguments{
\item{data}{Variable of class \code{\link{diallel_geno_pheno}}}

\item{trait}{Name of trait}

\item{params}{List containing burnIn and nIter}

\item{dominance}{Dominance degree (1-4)}

\item{chrom}{Names of chromosomes to scan (default is all)}

\item{cofactor}{Optional name of marker to include as cofactor in the scan}

\item{n.core}{Number of cores for parallel execution (only available from Linux or Mac command line)}
}
\value{
Data frame containing the map, LOD, r2 and deltaDIC results.
}
\description{
Performs a linear regression for each position in the map.
}
\details{
LOD score is the difference between the log10-likelihood for the QTL model vs. no QTL model (higher is better). deltaDIC is the difference between the Deviance Information Criterion for the QTL model vs. no QTL model (lower values is better). r2 is the squared correlation between the fitted and observed values. Parameter \code{dominance} controls the genetic model: 1 = additive, 2 = digenic dominance, 3 = trigenic dominance, 4 = quadrigenic dominance. MCMC \code{params} can be estimated using \code{\link{set_params}}.
}
\examples{
\dontrun{
  par1 <- set_params(data = diallel_example,
                     trait = "tuber_shape")
                      
  scan1_example <- scan1(data = diallel_example,
                chrom = 10,
                trait = "tuber_shape",
                params = par1)
}

}
