% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1_permute.R
\name{scan1_permute}
\alias{scan1_permute}
\title{Permutation test for scan1}
\usage{
scan1_permute(
  data,
  trait,
  params,
  n.permute = 1000,
  chrom = NULL,
  dominance = 1,
  covariate = NULL,
  n.core = 1
)
}
\arguments{
\item{data}{Variable of class \code{\link{diallel_geno_pheno}}}

\item{trait}{Name of trait}

\item{params}{List containing burnIn and nIter}

\item{n.permute}{Number of permutations}

\item{chrom}{Names of chromosomes to scan (default is all)}

\item{dominance}{Dominance degree (1-4)}

\item{covariate}{optional, to include markers as covariates. See Details.}

\item{n.core}{Number of cores for parallel execution}
}
\value{
Data frame with maximum LOD and minimum deltaDIC for each iteration
}
\description{
Permutation test for scan1
}
\examples{
\dontrun{
  set_params(data = diallel_example,
             trait = "tuber_shape")
                      
  ans1_permut <- scan1_permute(data = diallel_example,
                               chrom = 10,
                               trait = "tuber_shape",
                               params = list(burnIn=60,nIter=600),
                               n.permute = 100)
                               
  ## computing permutation threshold for alpha=0.05                            
  quantile(ans1_permut$min_deltaDIC, 0.05)
                             
}

}
