% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1_summary.R
\name{scan1_summary}
\alias{scan1_summary}
\title{Summary of scan1 result}
\usage{
scan1_summary(scan1_data, thresh = NULL, chrom = NULL, position = "cM")
}
\arguments{
\item{scan1_data}{output from scan1}

\item{thresh}{optional -deltaDIC threshold for plotting}

\item{chrom}{optional, subset of chromosomes to plot}

\item{position}{Either "cM" (default) or "bp"}
}
\value{
List containing
\describe{
\item{peaks}{data frame of markers with the highest score on each chromosome}
\item{plot}{ggplot object}
}
}
\description{
Summary of scan1 result
}
\details{
Plots the "score" (-deltaDIC) for each marker vs. genome position. The \code{thresh} argument should be a positive number.
}
\examples{
\dontrun{
  scan1_summary( scan1_example )
  scan1_summary( scan1_example, chrom = "10" )
  scan1_summary( scan1_example, chrom = c( "10", "12" ) ) 
  scan1_summary( scan1_example, chrom = "10", thresh = 20)
  }
}
