% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params.R
\name{set_params}
\alias{set_params}
\title{Determine burn-in and total number of iterations}
\usage{
set_params(
  data,
  trait,
  dominance = 1,
  marker = NULL,
  q = 0.5,
  r = 0.1,
  nIter = 2000
)
}
\arguments{
\item{data}{variable of class \code{\link{diallel_geno_pheno}}}

\item{trait}{name of trait}

\item{dominance}{dominance degree}

\item{marker}{name of marker (optional)}

\item{q}{quantile to estimate}

\item{r}{tolerance for quantile}

\item{nIter}{number of iterations}
}
\value{
List containing
\describe{
\item{burnIn}{Number of burn-in iterations}
\item{nIter}{Total number of iterations}
}
}
\description{
Determine burn-in and total number of iterations
}
\details{
Determines the burn-in and total number of iterations using the Raftery and Lewis diagnostic from R package \code{coda}, based on a 95\% probability that the estimate for quantile \code{q} of the additive effects is within the interval \code{(q-r,q+r)}. The 90th percentile for burn-in and total iterations across the additive effects is returned. Parameter \code{dominance} specifies which genetic model (1 = additive, 2 = digenic dominance, 3 = trigenic dominance, 4 = quadrigenic dominance) to use when determining the number of iterations, but this must be parameter must still be specified when calling functions such as \code{\link{scan1}} or \code{\link{fitQTL}}. Suggested values for \code{\link{scan1}} are q=0.5 and r=0.1. For \code{\link{fitQTL}}, the values depend on the desired Bayesican credible interval. For a 90\% CI, suggested values are q=0.05 and r=0.025. If \code{marker=NULL} (default), the first marker of every chromosome is analyzed to generate parameters suitable for \code{\link{scan1}}. Parameter \code{nIter} sets the number of iterations used to apply the Raftery and Lewis diagnostic; the default value is 2000, and if a larger number is needed, an error will be generated with this information.
}
\examples{
\dontrun{
  # Parameters for scan1
  par1 <- set_params(data = diallel_example,
                     trait = "tuber_shape",q=0.5,r=0.1)
                     
  # Parameters for fitQTL
  par2 <- set_params(data = diallel_example,
                     trait = "tuber_shape",q=0.05,r=0.05,marker="solcap_c2_25522")
}
  
}
